requirements.txt
# requirements.txt for Arepas
# Add required packages for OpenAI Vision, image processing, and data handling
openai
pandas
numpy
Pillow
loguru
